Update mixed precision handling in replay_critical_debug.py and training_lo_debug.py
- Modified the autocast context manager to explicitly use `dtype=torch.bfloat16` for improved numerical stability during mixed precision training on CUDA.
- Ensured consistency in mixed precision handling across both scripts, enhancing compatibility and performance during model evaluation.

+++ training_lo/replay_critical_debug.py
@@ -454,7 +454,7 @@ def main() -> None:
             handles, first_hit = register_nan_hooks(model, verbose=args.verbose_hooks)
         try:
             if args.mixed_precision and device.type == "cuda":
-                with autocast("cuda"):
+                with autocast("cuda", dtype=torch.bfloat16):
                     loss, info = run_elbo_inner()
             else:
                 loss, info = run_elbo_inner()
